Moved dependencywheel series into a subfolder.
