+++ js/Series/DependencyWheel/DependencyWheelSeries.js
@@ -23,13 +23,13 @@ var __extends = (this && this.__extends) || (function () {
         d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
     };
 })();
-import A from '../Core/Animation/AnimationUtilities.js';
+import A from '../../Core/Animation/AnimationUtilities.js';
 var animObject = A.animObject;
-import BaseSeries from '../Core/Series/Series.js';
+import BaseSeries from '../../Core/Series/Series.js';
 var _a = BaseSeries.seriesTypes, PieSeries = _a.pie, SankeySeries = _a.sankey;
-import H from '../Core/Globals.js';
-import NodesMixin from '../Mixins/Nodes.js';
-import U from '../Core/Utilities.js';
+import H from '../../Core/Globals.js';
+import NodesMixin from '../../Mixins/Nodes.js';
+import U from '../../Core/Utilities.js';
 var extend = U.extend, merge = U.merge;
 /**
  * @private
